moved mobile opt in scenario to a new file
